requirements.txt
aiogram==2.9
aiohttp==3.6.2
aioredis==1.3.1
asyncpg
python-dotenv
